match type refactor
