+++ bancho/recovery.js
@@ -1,5 +1,5 @@
 const { prisma } = require("../prisma");
-const { MatchManager } = require("./MatchManager");
+const { MatchManager } = require("./match-types/MatchManager");
 module.exports.recover = async () => {
 	let matches = await prisma.match.findMany({
 		where: {
